+++ nactionx/__init__.py
@@ -2,4 +2,4 @@
 
 APP_NAME = 'NactionX Downloader'
 APP_ID = 'NactionX.Downloader'
-__version__ = '1.2.1'
+__version__ = '1.2.2'